requirements.txt
# RAG Restaurant Management System
# Core Framework
fastapi>=0.104.0,<0.105.0
uvicorn[standard]>=0.24.0,<0.25.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# Authentication & Security
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.6,<1.0.0
PyJWT>=2.8.0,<3.0.0
bcrypt>=4.1.2,<5.0.0

# Database
sqlalchemy>=2.0.23,<3.0.0
psycopg2-binary>=2.9.9,<3.0.0
alembic>=1.13.1,<2.0.0

# Vector Database (Qdrant Cloud)
qdrant-client>=1.7.0,<2.0.0

# AI & ML
openai>=1.3.7,<2.0.0
numpy>=1.24.0,<2.0.0
scikit-learn>=1.3.0,<2.0.0
sentence-transformers>=2.2.0,<3.0.0

# LangChain & AI Framework
langchain>=0.1.0,<1.0.0
langchain-openai>=0.0.5,<1.0.0
langchain-community>=0.0.10,<1.0.0
langchain-core>=0.1.0,<1.0.0

# File Processing
PyMuPDF>=1.23.0,<2.0.0
python-magic>=0.4.27,<1.0.0
PyPDF2>=3.0.1,<4.0.0
python-docx>=1.1.0,<2.0.0
docx2txt>=0.8,<1.0.0  # DOC file text extraction
openpyxl>=3.1.0,<4.0.0
xlrd>=2.0.0,<3.0.0
Pillow>=10.0.0,<11.0.0
pandas>=2.0.0,<3.0.0
markdown>=3.5.0,<4.0.0
python-pptx>=0.6.0,<1.0.0
striprtf>=0.0.26,<1.0.0

# OCR Processing
pytesseract>=0.3.10,<1.0.0

# Async & Background Tasks
httpx>=0.25.0,<1.0.0

# Cache Cleanup & Scheduling
schedule>=1.2.0,<2.0.0

# Logging & Monitoring
structlog>=23.2.0,<24.0.0
python-json-logger==3.3.0

# Environment & Configuration
python-dotenv>=1.0.0,<2.0.0
click>=8.1.0,<9.0.0

# Testing & Development
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.21.0,<1.0.0

# Documentation
mkdocs>=1.5.0,<2.0.0
mkdocs-material>=9.4.0,<10.0.0

# Utilities
python-dateutil>=2.8.0,<3.0.0
pytz>=2023.0,<2024.0

# Cloud Services
supabase>=2.0.0,<3.0.0

# Security & Cryptography
cryptography>=41.0.0,<42.0.0

# Note: This project requires Python 3.8+ for compatibility
